add example to render 3d cube on 2d canvas with lines
and project every vertex by hand.

+++ class-01-basic-game-dev/04-0-render-basic/02_triangle_render_main.cxx
@@ -15,8 +15,9 @@ int main(int, char**)
 
     std::vector<position> triangle;
     triangle.push_back(position{ 0, 0 });
-    triangle.push_back(position{ width - 1, height - 1 });
-    triangle.push_back(position{ 0, height - 1 });
+    triangle.push_back(
+        position{ static_cast<int>(width - 1), static_cast<int>(height - 1) });
+    triangle.push_back(position{ 0, static_cast<int>(height - 1) });
 
     render_tri.draw_triangles(triangle, 3, green);
